deltaverse-api/requirements.txt
# Core API Dependencies (pinned for better caching)
#fastapi==0.116.1
fastapi>=0.104.1,<0.117.0
uvicorn[standard]==0.24.0
pydantic==2.11.7
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Firebase and Google Cloud (pinned versions)
firebase-admin==6.4.0
google-cloud-firestore==2.21.0
google-cloud-storage==2.19.0
google-cloud-pubsub==2.31.0
aiohttp==3.12.14

# Google AI and Vertex AI (pinned for stability)
google-generativeai==0.8.5
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0

# MCP Integration (pinned versions)
#mcp==1.12.0
mcp>=1.10.0,<1.13.0
httpx==0.28.1
httpx-sse==0.4.1
pydantic-settings==2.10.1
#anyio==4.9.0
anyio>=3.7.1





# Voice Recognition Dependencies (optional - can be removed if not used)
SpeechRecognition==3.14.3
pyaudio==0.2.14
websockets==15.0.1

# Additional dependencies (pinned)
requests==2.32.4
python-dotenv==1.0.0
tenacity==9.1.2
typing-extensions==4.14.1

# Testing (only needed for development)
pytest==8.4.1
pytest-asyncio==1.1.0

pydantic[email]